Refactor

+++ src/main/java/com/zpo/studentsystem/service/CourseService.java
@@ -13,6 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 
 @Service
@@ -22,47 +23,40 @@ public class CourseService {
     @Autowired
     private CourseRepository courseRepo;
 
-    @Autowired
-    private GradeRepository gradeRepo;
-
-    @Autowired
-    private StudentRepository studentRepo;
-
     public List<Course> getCourses() {
         return courseRepo.findAll();
     }
 
-
-    public void addCourse(String name) {
+    public Course addCourse(String name) {
         Course course = new Course();
         course.setName(name);
-        courseRepo.save(course);
+        return courseRepo.save(course);
     }
 
-    public void deleteCourse(Long id) {
+    public Boolean deleteCourse(Long id) {
+        if(getCourse(id) == null) return false;
         courseRepo.deleteById(id);
-    }
-
-    public void addGrade(Long studentId, Long courseId, Long points) {
-        System.out.println("Adding grade for student: " + studentId + " and course: " + courseId + " with points: " + points);
-        Student student = studentRepo.findById(studentId).orElseThrow(() -> new EntityNotFoundException("Student not found with id " + studentId));
-        System.out.println(student);
-        Course course = courseRepo.findById(courseId).orElseThrow(() -> new EntityNotFoundException("Course not found with id " + courseId));
-        System.out.println(course);
-        Grade grade = new Grade();
-        GradeId id = new GradeId();
-        id.setStudentId(studentId);
-        id.setCourseId(courseId);
-        grade.setId(id);
-        grade.setStudent(student);
-        grade.setCourse(course);
-        grade.setMaxPoints(10L);
-        grade.setPoints(points);
-        gradeRepo.save(grade);
+        return true;
     }
 
     public Course getCourse(Long courseId) {
         return courseRepo.findById(courseId).orElseThrow(() -> new EntityNotFoundException("Course not found with id " + courseId));
+    }
+
+    public HashMap<Course, Double> getAverageGrade() {
+        List<Course> courses = courseRepo.findAll();
+        courses.removeIf(it -> it.getGrades().isEmpty());
+        HashMap<Course, Double> averageGrades = new HashMap<>();
+        double value;
+
+        for(Course c : courses) {
+            value = c.getGrades().stream()
+                    .mapToDouble(Grade::getPoints)
+                    .average()
+                    .orElse(0.0);
+            averageGrades.put(c, value);
+        }
+        return averageGrades;
     }
 
 }
